requirements.txt
numpy
pandas
Pillow
torch
scikit-learn
torchvision
matplotlib
